requirements.txt
discord.py==2.3.2
pymongo==4.6.1
python-dotenv==1.0.1
Requests==2.31.0
